Move message kits to policy submodule
