+++ nucypher/crypto/keypairs.py
@@ -26,7 +26,6 @@ from hendrix.deploy.tls import HendrixDeployTLS
 from hendrix.facilities.services import ExistingKeyTLSContextFactory
 
 from nucypher.config.constants import MAX_UPLOAD_CONTENT_LENGTH
-from nucypher.crypto.kits import MessageKit
 from nucypher.crypto.signing import SignatureStamp, StrangerStamp
 from nucypher.crypto.tls import _read_tls_certificate, _TLS_CURVE, generate_self_signed_certificate
 from nucypher.crypto.umbral_adapter import (
@@ -38,6 +37,7 @@ from nucypher.crypto.umbral_adapter import (
     decrypt_reencrypted
 )
 from nucypher.network.resources import get_static_resources
+from nucypher.policy.kits import MessageKit
 
 
 class Keypair(object):
