+++ tests/acceptance/characters/test_decentralized_grant.py
@@ -21,9 +21,9 @@ import datetime
 import maya
 import pytest
 
-from nucypher.crypto.kits import MessageKit
 from nucypher.crypto.utils import keccak_digest
 from nucypher.datastore.models import EncryptedTreasureMap as DatastoreTreasureMap
+from nucypher.policy.kits import MessageKit
 from nucypher.policy.maps import EncryptedTreasureMap
