+++ tests/integration/characters/control/test_web_control_federated.py
@@ -25,8 +25,8 @@ import pytest
 from click.testing import CliRunner
 
 import nucypher
-from nucypher.crypto.kits import MessageKit
 from nucypher.crypto.powers import DecryptingPower
+from nucypher.policy.kits import MessageKit
 from nucypher.policy.maps import EncryptedTreasureMap
 
 click_runner = CliRunner()
