+++ tests/unit/characters/control/test_character_fields.py
@@ -20,9 +20,6 @@ from base64 import b64encode, b64decode
 import maya
 import pytest
 from nucypher.crypto.umbral_adapter import SecretKey, Signer
-
-from nucypher.crypto.kits import MessageKit as MessageKitClass
-
 from nucypher.characters.control.specifications.fields import (
     DateTime,
     FileField,
@@ -33,6 +30,7 @@ from nucypher.characters.control.specifications.fields import (
 )
 from nucypher.characters.lawful import Enrico
 from nucypher.control.specifications.exceptions import InvalidInputData
+from nucypher.policy.kits import MessageKit as MessageKitClass
 
 
 #
